Added generic dialog component

+++ src/components/profile/venue-list-item/index.tsx
@@ -1,25 +1,15 @@
 import { ReactElement, useState } from 'react'
 import { Link } from 'react-router-dom'
-import {
-    Avatar,
-    Box,
-    Button,
-    Dialog,
-    DialogActions,
-    DialogContent,
-    DialogContentText,
-    DialogTitle,
-    IconButton,
-    ListItem,
-    ListItemAvatar,
-    Tooltip,
-} from '@mui/material'
+import { useDeleteVenueMutation } from '@/services/holidaze.ts'
 
+import { Avatar, Box, Button, IconButton, ListItem, ListItemAvatar, Tooltip } from '@mui/material'
 import DeleteIcon from '@mui/icons-material/Delete'
 import ListItemText from '@mui/material/ListItemText'
+
 import VenuePlaceholder from '@/assets/venue-placeholder.svg'
+import AlertDialog from '@/components/common/dialog'
+
 import { VenueLocation } from '@/types/venue.ts'
-import { useDeleteVenueMutation } from '@/services/holidaze.ts'
 
 interface VenueCardProps {
     id: string
@@ -60,12 +50,17 @@ export default function VenueListItem(props: VenueCardProps): ReactElement {
             secondaryAction={
                 <Box>
                     <Tooltip title={'Delete Venue'}>
-                        <IconButton aria-label={`Delete venue ${name}`} onClick={handleClickOpen} color="error">
+                        <IconButton aria-label={`Delete venue ${name}`} onClick={handleClickOpen}>
                             <DeleteIcon />
                         </IconButton>
                     </Tooltip>
-                    <Button component={Link} to={`/venues/admin/${id}`} aria-label={`View ${name}`}>
-                        Venue Admin
+                    <Button
+                        variant="outlined"
+                        component={Link}
+                        to={`/venues/${id}`}
+                        aria-label={`View and administer ${name} venue`}
+                    >
+                        View
                     </Button>
                 </Box>
             }
@@ -74,32 +69,16 @@ export default function VenueListItem(props: VenueCardProps): ReactElement {
                 <Avatar src={src} alt="" />
             </ListItemAvatar>
             <ListItemText primary={name} secondary={secondary} />
-            <Dialog
+            <AlertDialog
                 open={open}
-                onClose={handleClose}
-                aria-labelledby="alert-dialog-title"
-                aria-describedby="alert-dialog-description"
-            >
-                <DialogTitle id="alert-dialog-title">{'Are you sure you want to delete the venue?'}</DialogTitle>
-                <DialogContent>
-                    <DialogContentText id="alert-dialog-description">
-                        Are you sure you want to delete the venue named {name}? This action can not be undone.
-                    </DialogContentText>
-                </DialogContent>
-                <DialogActions>
-                    <Button onClick={handleClose}>Cancel</Button>
-                    <Button
-                        variant="contained"
-                        color="error"
-                        onClick={async (): Promise<void> => {
-                            await handleDelete(id)
-                        }}
-                        autoFocus
-                    >
-                        Delete
-                    </Button>
-                </DialogActions>
-            </Dialog>
+                title={`Delete ${name}?`}
+                text={`Are you sure you want to delete the venue named ${name}? This action can not be undone.`}
+                primaryColor={'error'}
+                action={async (): Promise<void> => {
+                    await handleDelete(id)
+                }}
+                handleClose={handleClose}
+            />
         </ListItem>
     )
 }
